AIMS API Key retriever

This function tries to search for an API Key

find_api_key(api_key)

Arguments

api_key
An API Key obtained from AIMS DataPlatform.

Value

Either a character vector API Key found in .Renviron or, if missing entirely, an error message.

Details

The AIMS Data Platform R Client provides easy access to data sets for R applications to the AIMS Data Platform API. The AIMS Data Platform requires an API Key for requests, which can be obtained at this link. It is preferred that API Keys are not stored in code. We recommend storing the environment variable AIMS_DATAPLATFORM_API_KEY permanently under the user's .Renviron file in order to load the API Key automatically.

Author

AIMS Datacentre adc@aims.gov.au