Format json_results output

Wrapper function

process_request(dt_req, next_page = FALSE, ...)

Arguments

dt_req

An URL GET output

next_page

Logical. Is this a multi-url request?

...

Additional arguments to be passed to internal function update_format

Value

aims_data returns a data.frame of class aimsdf.

If summary %in% c("summary-by-series", "summary-by-deployment"), the output shows the summary information for the target dataset (i.e. weather or temperature loggers) (NB: currently, summary only works for the temperature logger database). If summary is not passed as an additional argument, then the output contains raw monitoring data. If summary = "daily", then the output contains mean daily aggregated monitoring data. The output also contains five attributes (empty strings if summary is passed as an additional argument):

metadataa DOI link containing the metadata record for the data series.

citationthe citation information for the particular dataset.

parametersThe measured parameters comprised in the output.

typeThe type of dataset. Either "monitoring" if summary is not specified, "monitoring (daily aggregation)" if summary = "daily", or a "summary-by-" otherwise.

targetThe input target.

Details

This function checks for errors in dt_req data request and processes result via json_results.

Author

AIMS Datacentre adc@aims.gov.au